+++ torchao/_models/llama/eval.py
@@ -31,7 +31,7 @@ from torchao._models.llama.model import prepare_inputs_for_model
 from tokenizer import get_tokenizer
 import time
 from torchao.utils import TORCH_VERSION_AT_LEAST_2_5
-from torchao.quantization.spin_quant import apply_spinquant
+from torchao.prototype.spinquant.spinquant import apply_spinquant
 
 def run_evaluation(
     checkpoint_path: Path,
